+++ berkeley-function-call-leaderboard/bfcl/model_handler/constant.py
@@ -112,7 +112,8 @@ JS_TYPE_CONVERSION = {
 }
 
 UNDERSCORE_TO_DOT = [
-    "gpt-4o-2024-08-06-FC",
+    # TODO: Use the model style to determine this, single source of truth
+    "gpt-4o-2024-11-20-FC",
     "gpt-4o-2024-05-13-FC",
     "gpt-4o-mini-2024-07-18-FC",
     "gpt-4-turbo-2024-04-09-FC",
@@ -126,6 +127,9 @@ UNDERSCORE_TO_DOT = [
     "claude-3-5-sonnet-20241022-FC",
     "claude-3-haiku-20240307-FC",
     "claude-3-5-haiku-20241022-FC",
+    "nova-pro-v1.0",
+    "nova-lite-v1.0",
+    "nova-micro-v1.0",
     "open-mistral-nemo-2407-FC",
     "open-mixtral-8x22b-FC",
     "mistral-large-2407-FC",
@@ -151,4 +155,5 @@ UNDERSCORE_TO_DOT = [
     "ibm-granite/granite-20b-functioncalling",
     "yi-large-fc",
     "openbmb/MiniCPM3-4B-FC",
+    "grok-beta",
 ]