+++ berkeley-function-call-leaderboard/bfcl/model_handler/handler_map.py
@@ -21,8 +21,10 @@ from bfcl.model_handler.proprietary_model.functionary import FunctionaryHandler
 from bfcl.model_handler.proprietary_model.gemini import GeminiHandler
 from bfcl.model_handler.proprietary_model.gogoagent import GoGoAgentHandler
 from bfcl.model_handler.proprietary_model.gorilla import GorillaHandler
+from bfcl.model_handler.proprietary_model.grok import GrokHandler
 from bfcl.model_handler.proprietary_model.mistral import MistralHandler
 from bfcl.model_handler.proprietary_model.nexus import NexusHandler
+from bfcl.model_handler.proprietary_model.nova import NovaHandler
 from bfcl.model_handler.proprietary_model.nvidia import NvidiaHandler
 from bfcl.model_handler.proprietary_model.openai import OpenAIHandler
 from bfcl.model_handler.proprietary_model.writer import WriterHandler
@@ -35,8 +37,8 @@ api_inference_handler_map = {
     "gorilla-openfunctions-v2": GorillaHandler,
     "o1-preview-2024-09-12": OpenAIHandler,
     "o1-mini-2024-09-12": OpenAIHandler,
-    "gpt-4o-2024-08-06": OpenAIHandler,
-    "gpt-4o-2024-08-06-FC": OpenAIHandler,
+    "gpt-4o-2024-11-20": OpenAIHandler,
+    "gpt-4o-2024-11-20-FC": OpenAIHandler,
     "gpt-4o-mini-2024-07-18": OpenAIHandler,
     "gpt-4o-mini-2024-07-18-FC": OpenAIHandler,
     "gpt-4-turbo-2024-04-09": OpenAIHandler,
@@ -45,16 +47,13 @@ api_inference_handler_map = {
     "gpt-3.5-turbo-0125-FC": OpenAIHandler,
     "claude-3-opus-20240229": ClaudeHandler,
     "claude-3-opus-20240229-FC": ClaudeHandler,
-    "claude-3-sonnet-20240229": ClaudeHandler,
-    "claude-3-sonnet-20240229-FC": ClaudeHandler,
-    "claude-3-5-sonnet-20240620": ClaudeHandler,
-    "claude-3-5-sonnet-20240620-FC": ClaudeHandler,
     "claude-3-5-sonnet-20241022": ClaudeHandler,
     "claude-3-5-sonnet-20241022-FC": ClaudeHandler,
-    "claude-3-haiku-20240307": ClaudeHandler,
-    "claude-3-haiku-20240307-FC": ClaudeHandler,
     "claude-3-5-haiku-20241022": ClaudeHandler,
     "claude-3-5-haiku-20241022-FC": ClaudeHandler,
+    "nova-pro-v1.0": NovaHandler,
+    "nova-lite-v1.0": NovaHandler,
+    "nova-micro-v1.0": NovaHandler,
     "open-mistral-nemo-2407": MistralHandler,
     "open-mistral-nemo-2407-FC": MistralHandler,
     "open-mixtral-8x22b": MistralHandler,
@@ -90,6 +89,7 @@ api_inference_handler_map = {
     "BitAgent/GoGoAgent": GoGoAgentHandler,
     # "yi-large-fc": YiHandler,  #  Their API is under maintenance, and will not be back online in the near future
     "palmyra-x-004": WriterHandler,
+    "grok-beta": GrokHandler,
 }
 
 # Inference through local hosting
@@ -123,10 +123,10 @@ local_inference_handler_map = {
     "NousResearch/Hermes-2-Pro-Llama-3-70B": HermesHandler,
     "NousResearch/Hermes-2-Theta-Llama-3-70B": HermesHandler,
     "ibm-granite/granite-20b-functioncalling": GraniteHandler,
-    "MadeAgents/Hammer2.0-7b": HammerHandler,
-    "MadeAgents/Hammer2.0-3b": HammerHandler,
-    "MadeAgents/Hammer2.0-1.5b": HammerHandler,
-    "MadeAgents/Hammer2.0-0.5b": HammerHandler,
+    "MadeAgents/Hammer2.1-7b": HammerHandler,
+    "MadeAgents/Hammer2.1-3b": HammerHandler,
+    "MadeAgents/Hammer2.1-1.5b": HammerHandler,
+    "MadeAgents/Hammer2.1-0.5b": HammerHandler,
     "THUDM/glm-4-9b-chat": GLMHandler,
     "Qwen/Qwen2-1.5B-Instruct": QwenHandler,
     "Qwen/Qwen2-7B-Instruct": QwenHandler,
@@ -146,6 +146,8 @@ local_inference_handler_map = {
 # Deprecated/outdated models, no longer on the leaderboard
 outdated_model_handler_map = {
     # "gorilla-openfunctions-v0": GorillaHandler,
+    # "gpt-4o-2024-08-06": OpenAIHandler,
+    # "gpt-4o-2024-08-06-FC": OpenAIHandler,
     # "gpt-4o-2024-05-13": OpenAIHandler,
     # "gpt-4o-2024-05-13-FC": OpenAIHandler,
     # "gpt-4-1106-preview-FC": OpenAIHandler,
@@ -156,6 +158,12 @@ outdated_model_handler_map = {
     # "gpt-4-0613": OpenAIHandler,
     # "claude-2.1": ClaudeHandler,
     # "claude-instant-1.2": ClaudeHandler,
+    # "claude-3-sonnet-20240229": ClaudeHandler,
+    # "claude-3-sonnet-20240229-FC": ClaudeHandler,
+    # "claude-3-5-sonnet-20240620": ClaudeHandler,
+    # "claude-3-5-sonnet-20240620-FC": ClaudeHandler,
+    # "claude-3-haiku-20240307": ClaudeHandler,
+    # "claude-3-haiku-20240307-FC": ClaudeHandler,
     # "gemini-1.0-pro-001": GeminiHandler,
     # "gemini-1.0-pro-001-FC": GeminiHandler,
     # "meetkai/functionary-small-v3.1-FC": FunctionaryHandler,